Update README.md
Recently, the detection of defects in raw materials became one of the important topics in deep learning. However, the lack of databases restricted the ability to build new models, since the improvement of performance requires more data. In this paper, the most recent database of strip steel defects is used to build a new technique for defects detection. Different pre-trained models were utilized for the feature extraction stage. Then, a deep neural network classifier was devised and trained on images for seven types of steel defects. The final model was able to classify the defects with a 98.43% accuracy, which was trained on the features extracted from the DesneNet201 pre-trained model. The results of the proposed framework outperformed the most recent work with a 2.93% of accuracy improvement.